Simplify fading-related calculations

+++ clownaudio/mixer.c
@@ -223,7 +223,7 @@ void Mixer_FadeOutSound(Mixer_Sound instance, unsigned int duration)
 	if (channel != NULL)
 	{
 		channel->fade_out_counter_max = (output_sample_rate * duration) / 1000;
-		channel->fade_counter = channel->fade_in_counter_max ? (((unsigned long long)channel->fade_in_counter_max - channel->fade_counter) * channel->fade_out_counter_max) / channel->fade_in_counter_max : channel->fade_out_counter_max;
+		channel->fade_counter = channel->fade_in_counter_max ? (channel->fade_in_counter_max - channel->fade_counter) * ((float)channel->fade_out_counter_max / (float)channel->fade_in_counter_max) : channel->fade_out_counter_max;
 		channel->fade_in_counter_max = 0;
 	}
 
@@ -239,7 +239,7 @@ void Mixer_FadeInSound(Mixer_Sound instance, unsigned int duration)
 	if (channel != NULL)
 	{
 		channel->fade_in_counter_max = (output_sample_rate * duration) / 1000;
-		channel->fade_counter = channel->fade_out_counter_max ? (((unsigned long long)channel->fade_out_counter_max - channel->fade_counter) * channel->fade_in_counter_max) / channel->fade_out_counter_max : channel->fade_in_counter_max;
+		channel->fade_counter = channel->fade_out_counter_max ? (channel->fade_out_counter_max - channel->fade_counter) * ((float)channel->fade_in_counter_max / (float)channel->fade_out_counter_max) : channel->fade_in_counter_max;
 		channel->fade_out_counter_max = 0;
 	}
